+++ src/App.jsx
@@ -14,8 +14,8 @@ import Mypage from './pages/Mypage';
 import Details from './pages/Details';
 import Voice from './pages/Voice';
 import Onboarding from './pages/Onboarding';
-import Mymain from './pages/Mymain';
 import Todayprofile from './pages/Todayprofile';
+import Monthly from './pages/Monthly';
 
 function App() {
   return (
@@ -24,6 +24,7 @@ function App() {
       <Route path="/signup" element={<Signup />} />
       <Route path="/signup/profile" element={<Profile />} />
       <Route path="/onboarding" element={<Onboarding />} />
+      <Route path="/onboarding/today" element={<Todayprofile />} />
       <Route path="/login" element={<Login />} />
       <Route path="/home" element={<Home />} />
       <Route path="/write" element={<Write />} />
@@ -33,8 +34,7 @@ function App() {
       <Route path="/recommend/result" element={<Result />} />
       <Route path="/chat" element={<Chat />} />
       <Route path="/mypage" element={<Mypage />} />
-      <Route path="/mymain" element={<Mymain />} />
-      <Route path="/todayprofile" element={<Todayprofile />} />
+      <Route path="/mypage/monthly" element={<Monthly />} />
       <Route path="/diary/:diaryDate" element={<Details />} />
       <Route path="/voice" element={<Voice />} />
     </Routes>
